Removing Jersey and fixing some critical bugs
Closing this set of develop by ripping out Jersey DI as it was not playing nicely with Spring.

Fixes some Jackson processing quirks as well with Cryptowat.ch sneaking in additional fields. It's not super clean but it will get the quote listing working (with duplicates).

See you on the other side.